Optimisation de l'interface administrateur avec l'ajout de plusieurs nouvelles fonctionnalités : bouton pour changer la langue, option pour supprimer des éléments, bouton de déconnexion et messages d'erreur intégrés.

+++ src/Controller/Admin/SpaceAmenitiesController.php
@@ -2,26 +2,43 @@
 
 namespace App\Controller\Admin;
 
+use Exception;
 use App\Entity\SpaceAmenities;
-use App\Form\SpaceAmenitiesType;
-use App\Repository\SpaceAmenitiesRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Contracts\Translation\TranslatorInterface;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/space/amenities')]
 class SpaceAmenitiesController extends AbstractController
 {
     #[Route('/{id}', name: 'app_space_amenities_delete', methods: ['POST'])]
-    public function delete(Request $request, SpaceAmenities $spaceAmenity, EntityManagerInterface $entityManager): Response
-    {
+    public function delete(
+        Request $request, 
+        SpaceAmenities $spaceAmenity, 
+        EntityManagerInterface $entityManager,
+        SessionInterface $session,
+        TranslatorInterface $trans,
+    ): Response {
+        $currentUser = $this->getUser();
+        if (!in_array('ROLE_ADMIN', $currentUser->getRoles())) return $this->redirectToRoute('public_home');
+        
         if ($this->isCsrfTokenValid('delete'.$spaceAmenity->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($spaceAmenity);
-            $entityManager->flush();
+            try {
+                $entityManager->remove($spaceAmenity);
+                $entityManager->flush();
+
+                $flashMessage = $trans->trans("The space type $spaceAmenity has been deleted.");
+                $session->getFlashBag()->add('success', $flashMessage);
+            } catch (Exception $e) {
+                $flashMessage = $trans->trans("An error has occurred.");
+                $session->getFlashBag()->add('error', $flashMessage);
+            }
         }
 
-        return $this->redirectToRoute('app_admin', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('admin_analytics', [], Response::HTTP_SEE_OTHER);
     }
 }
